+++ test/utils/structs/EnumerableMap.behavior.js
@@ -36,6 +36,13 @@ function shouldBehaveLikeMap (
     }))).to.have.same.deep.members(
       zip(keys.map(k => k.toString()), values.map(v => v.toString())),
     );
+
+    // This also checks that both arrays have the same length
+    expect(
+      (await methods.keys(map)).map(k => k.toString()),
+    ).to.have.same.members(
+      keys.map(key => key.toString()),
+    );
   }
 
   it('starts empty', async function () {